+++ src/collective/iconifiedcategory/browser/tabview.py
@@ -266,7 +266,7 @@ class FilesizeColumn(BaseColumn):
 
 
 class IconClickableColumn(BaseColumn):
-    action_view = ''
+    action_view_name = ''
     escape = False
 
     def _deactivated_is_useable(self):
@@ -282,7 +282,21 @@ class IconClickableColumn(BaseColumn):
         )
 
     def get_action_view(self, content):
-        return self.action_view
+        if content.UID in getattr(self, '_cached_action_views', {}):
+            view = self._cached_action_views[content.UID]
+        else:
+            view = getMultiAdapter((content.getObject(), self.request), name=self.action_view_name)
+            self._store_cached_view(content.UID, view)
+        return view
+
+    def get_action_view_name(self, content):
+        return self.action_view_name
+
+    def _store_cached_view(self, value, view):
+        """ """
+        if getattr(self, '_cached_action_views', None) is None:
+            self._cached_action_views = {}
+        self._cached_action_views[value] = view
 
     def alt(self, content):
         return self.header
@@ -291,7 +305,7 @@ class IconClickableColumn(BaseColumn):
         return getattr(content, self.attrName, False) is None
 
     def is_editable(self, content):
-        view = getMultiAdapter((content.getObject(), self.request), name=self.action_view)
+        view = self.get_action_view(content)
         return view._may_set_values({})
 
     def is_active(self, content):
@@ -323,7 +337,7 @@ class PrintColumn(IconClickableColumn):
     cssClasses = {'td': 'iconified-print'}
     weight = 80
     attrName = 'to_print'
-    action_view = 'iconified-print'
+    action_view_name = 'iconified-print'
 
     def alt(self, content):
         return translate(
@@ -338,7 +352,7 @@ class ConfidentialColumn(IconClickableColumn):
     cssClasses = {'td': 'iconified-confidential'}
     weight = 90
     attrName = 'confidential'
-    action_view = 'iconified-confidential'
+    action_view_name = 'iconified-confidential'
 
     def alt(self, content):
         return translate(
@@ -353,7 +367,7 @@ class SignedColumn(IconClickableColumn):
     cssClasses = {'td': 'iconified-signed'}
     weight = 95
     attrName = 'signed'
-    action_view = 'iconified-signed'
+    action_view_name = 'iconified-signed'
 
     def alt(self, content):
         return translate(
@@ -375,7 +389,7 @@ class ApprovedColumn(IconClickableColumn):
     cssClasses = {'td': 'iconified-approved'}
     weight = 100
     attrName = 'approved'
-    action_view = 'iconified-approved'
+    action_view_name = 'iconified-approved'
 
     def alt(self, content):
         return translate(
@@ -397,7 +411,7 @@ class PublishableColumn(IconClickableColumn):
     cssClasses = {'td': 'iconified-publishable'}
     weight = 98
     attrName = 'publishable'
-    action_view = 'iconified-publishable'
+    action_view_name = 'iconified-publishable'
 
     def alt(self, content):
         return translate(